# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "CHELSY Restaurant API",
        "description": "API REST pour l'application mobile CHELSY Restaurant",
        "contact": {
            "email": "contact@chelsy-restaurant.bj"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://my-default-host.com",
            "description": "Serveur API"
        }
    ],
    "paths": {
...
        "/api/v1/orders/{order_id}/tracking": {
            "get": {
                "tags": [
                    "Suivi GPS Livreur"
                ],
                "summary": "Récupérer la position du livreur pour une commande (client)",
                "operationId": "93344a6a8f47b04d85af21da2c4e2175",
                "parameters": [
                    {
                        "name": "order_id",
                        "in": "path",
                        "description": "ID de la commande",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Position du livreur",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "properties": {
                                                "position": {
                                                    "type": "object",
                                                    "nullable": true
                                                },
                                                "driver": {
                                                    "type": "object",
                                                    "nullable": true
                                                },
                                                "eta_minutes": {
                                                    "description": "Temps estimé d'arrivée en minutes",
                                                    "type": "integer",
                                                    "nullable": true
                                                },
                                                "distance_km": {
                                                    "description": "Distance en kilomètres",
                                                    "type": "number",
                                                    "format": "float",
                                                    "nullable": true
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Commande n'appartient pas à l'utilisateur"
                    },
                    "404": {
                        "description": "Commande non trouvée"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Authentification par token Sanctum",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}